Improve support for SMPTE-style timing
This patch fixes a bug with the decoding of the SMPTE divisions
field.

As an example, the bug would cause the commonly-used "E7 28" to
be decoded to 960 ticks per beat, instead of the desired 1000
ticks per second.

This was caused by an error in the bit-twiddling integer decoding,
and also by failing to divide by two before interpreting ticks
per second as ticks per beat (at the default 120 BPM).